# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/space-options:
    get:
      operationId: listSpaceOptions
      tags: [space-options]
      summary: List space options
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of space options
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpaceOption"
    post:
      operationId: createSpaceOption
      tags: [space-options]
      summary: Create space option
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSpaceOptionBody"
      responses:
        "201":
          description: Created space option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceOption"
  /v1/space-options/{id}:
    get:
      operationId: getSpaceOption
      tags: [space-options]
      summary: Get space option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Space option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceOption"
        "404":
          description: Not found
    put:
      operationId: updateSpaceOption
      tags: [space-options]
      summary: Update space option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpaceOptionBody"
      responses:
        "200":
          description: Updated space option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceOption"
        "404":
          description: Not found
    delete:
      operationId: deleteSpaceOption
      tags: [space-options]
      summary: Delete space option
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─────────────────────────────────────────────
  # SPACE POLICIES
  # ─────────────────────────────────────────────
...
  /v1/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: lead_status
          in: query
          schema:
            type: string
        - name: lead_source
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: nationality
          in: query
          schema:
            type: string
        - name: preferred_space_type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /v1/leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete (soft) lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /v1/leads/{id}/convert:
    patch:
      operationId: convertLead
      tags: [leads]
      summary: Convert lead to booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertLeadBody"
      responses:
        "200":
          description: Conversion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id:
                    type: integer
                  booking_ref:
                    type: string
                  lead_ref:
                    type: string
                required: [booking_ref, lead_ref]
  /v1/leads/{id}/mark-lost:
    patch:
      operationId: markLeadLost
      tags: [leads]
      summary: Mark lead as lost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
components:
  schemas:
...
    SpaceOption:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        display_name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - created_at
        - updated_at
    CreateSpaceOptionBody:
      type: object
      properties:
        name:
          type: string
        display_name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string
          default: Active
      required:
        - name
    UpdateSpaceOptionBody:
      type: object
      properties:
        name:
          type: string
        display_name:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        status:
          type: string

    # ─── SPACE POLICY ───
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        lead_ref:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        lead_source:
          type: ["string", "null"]
        lead_status:
          type: string
        inquiry_type:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        preferred_space_type:
          type: ["string", "null"]
        preferred_check_in_date:
          type: ["string", "null"]
        preferred_duration_weeks:
          type: ["integer", "null"]
        preferred_suburb_id:
          type: ["integer", "null"]
        budget_min:
          type: ["string", "null"]
        budget_max:
          type: ["string", "null"]
        budget_currency:
          type: ["string", "null"]
        converted_booking_id:
          type: ["integer", "null"]
        converted_at:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        manual_input:
          type: boolean
        status:
          type: string
        preferred_suburb_name:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, lead_ref, first_name, last_name, email, lead_status, manual_input, status, created_at, updated_at]
    CreateLeadBody:
      type: object
      properties:
        lead_ref:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        lead_source:
          type: ["string", "null"]
        lead_status:
          type: string
          default: New
        inquiry_type:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        preferred_space_type:
          type: ["string", "null"]
        preferred_check_in_date:
          type: ["string", "null"]
        preferred_duration_weeks:
          type: ["integer", "null"]
        preferred_suburb_id:
          type: ["integer", "null"]
        budget_min:
          type: ["string", "null"]
        budget_max:
          type: ["string", "null"]
        budget_currency:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        manual_input:
          type: boolean
          default: false
        status:
          type: string
          default: Active
      required: [first_name, last_name, email]
    UpdateLeadBody:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        lead_source:
          type: ["string", "null"]
        lead_status:
          type: string
        inquiry_type:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        preferred_space_type:
          type: ["string", "null"]
        preferred_check_in_date:
          type: ["string", "null"]
        preferred_duration_weeks:
          type: ["integer", "null"]
        preferred_suburb_id:
          type: ["integer", "null"]
        budget_min:
          type: ["string", "null"]
        budget_max:
          type: ["string", "null"]
        budget_currency:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        manual_input:
          type: boolean
        status:
          type: string
...
    ConvertLeadBody:
      type: object
      properties:
        space_id:
          type: integer
        check_in_date:
          type: string
        check_out_date:
          type: string
        agreed_weekly_rate:
          type: ["string", "null"]
      required: [space_id, check_in_date, check_out_date]
...